Distribute tests and implementation into more appropriate files
